+++ src/bouncer/helpers.clj
@@ -1,4 +1,5 @@
-(ns bouncer.helpers)
+(ns bouncer.helpers
+  {:author "Leonardo Borges"})
 
 (defn resolve-or-same [form]
   (if (and (symbol? form)