+++ src/bouncer/validators.clj
@@ -1,15 +1,24 @@
 (ns bouncer.validators
+  "This namespace contains all built-in validators as well as
+          macros for defining new validators and validator sets"
+  {:author "Leonardo Borges"}
   (:require [clojure.walk :as w]
             [bouncer.helpers :as h]))
+
+;; ## Customization support
+;;
+;; The following functions and macros support creating custom validators
 
 (defn mk-validator
   "Returns a validation function that will use (pred (k m)) to determine if m is valid. msg will be added to the :errors entry in invalid scenarios.
 
-If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
+  If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
 
-A validator can also be marked optional, in which case the validation will only run if k has a value in m. e.g.:
+  A validator can also be marked optional, in which case the validation will only run if k has a value in m.
 
-  (mk-validator number? k msg :optional true)"
+  e.g.:
+
+    (mk-validator number? k msg :optional true)"
   ([pred k msg & {optional :optional}]
      (fn [m]
        (if (and optional (cond
@@ -40,8 +49,30 @@ A validator can also be marked optional, in which case the validation will only 
   docstring and opts-map are optional
 
   opts-map is a map of key-value pairs and may be one of:
-    :default-message-format used when the client of this validator doesn't provide a message
-    :optional whether the validation should be run only if the given key has a non-nil value in the map. Defaults to false.
+
+  `:default-message-format` used when the client of this validator doesn't provide a message
+
+  `:optional` whether the validation should be run only if the given key has a non-nil value in the map. Defaults to false.
+
+  The function will be called with the value being validated as its first argument.
+
+  Any extra arguments will be passed along to the function in the order they were used in the 
+  \"validate\" call.
+
+  e.g.:
+
+
+    (defvalidator in
+      [value coll]
+      (some #{value} coll))
+
+    (validate {:age 10}
+      :age (in (range 5)))
+
+
+  This means the validator `in` will be called with the arguments `10` and `(0 1 2 3 4)`, 
+  in that order.
+
 "
   {:arglists '([name docstring? opts-map? [args*] & body])}
   [name & options]
@@ -57,31 +88,29 @@ A validator can also be marked optional, in which case the validation will only 
         options (if (map? (first options))
                   (next options)
                   options)
-        [args & body] options]
+        [args & body] options
+        [pred-subject & rest] args]
     `(defn ~(with-meta name {:doc docstring})
-       {:arglists '([~(symbol 'k) ~(symbol 'opts?)])}
-       ([k#]
-          (~name k# :message (format (if ~default-message-format
-                                       ~default-message-format
-                                       "Custom validation failed for %s") (key->name k#))))
-       ([k# & {message# :message}]
-          (mk-validator (fn ~args
-                          ~@body) k# message# :optional ~optional)))))
+       {:arglists '([~@args])}
+       ([k# ~@rest]
+          (~name k# ~@rest :message (format (if ~default-message-format
+                                              ~default-message-format
+                                              "Custom validation failed for %s") (key->name k#))))
+       ([k# ~@rest & {message# :message}]
+          (mk-validator (fn [~pred-subject]
+                          (let [~@(mapcat #(vector %1 %2) rest rest)]
+                            ~@body))
+                        k# message# :optional ~optional)))))
 
-;; Default validators
+
+;; ## Built-in validators
+
 (defvalidator required
-  "Validating function that checks if its only argument has a value.
+  "Validates value is present.
 
   If the value is a string, it makes sure it's not empty, otherwise it checks for nils.
 
-  The resulting function takes a key k.
-
-  If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
-
-  opts is an optional sequence of named arguments and may be one of:
-    - :message the message to be added to the :errors map should this validation fail
-  
-  If no message is given, a default message will be used
+  For use with validation macros such as `validate` or `valid?`
 "
   {:default-message-format "%s must be present"}
   [value]
@@ -90,80 +119,48 @@ A validator can also be marked optional, in which case the validation will only 
     (not (nil? value))))
 
 (defvalidator number
-  " Validating function that checks its only argument is a number.
+  "Validates maybe-a-number is a valid number.
 
-  The resulting function takes a key k.
-
-  If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
-
-  opts is an optional sequence of named arguments and may be one of:
-    - :message the message to be added to the :errors map should this validation fail
-  
-  If no message is given, a default message will be used
-"
+  For use with validation macros such as `validate` or `valid?`"
   {:default-message-format "%s must be a number" :optional true}
   [maybe-a-number]
   (number? maybe-a-number))
 
 
 (defvalidator positive
-  " Validating function that checks its only argument is greater than zero.
+  "Validates number is a number and is greater than zero.
 
-  The resulting function takes a key k.
-
-  If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
-
-  opts is an optional sequence of named arguments and may be one of:
-    - :message the message to be added to the :errors map should this validation fail
-  
-  If no message is given, a default message will be used
-"
+  For use with validation macros such as `validate` or `valid?`"
   {:default-message-format "%s must be a positive number" :optional true}
   [number]
   (and (number? number)
        (> number 0)))
 
 
+(defvalidator member
+  "Validates value is a member of coll.
+
+  For use with validation macros such as `validate` or `valid?`"
+  {:default-message-format "%s out of range"}
+  [value coll]
+  (some #{value} coll))
+
 (defvalidator custom
-  " Validating function that checks its only argument is greater than zero.
+  "Validates pred is true for the given value.
 
-  The resulting function takes a key k.
+  For use with validation macros such as `validate` or `valid?`"
+  [value pred]
+  (pred value))
 
-  If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
+(defvalidator every
+  "Validates pred is true for every item in coll.
 
-  opts is an optional sequence of named arguments and may be one of:
-    - :message the message to be added to the :errors map should this validation fail
-  
-  If no message is given, a default message will be used
-"
-  [])
+  For use with validation macros such as `validate` or `valid?`"
+  {:default-message-format "All items in %s must satisfy the predicate"}
+  [coll pred]
+  (every? pred coll)) 
 
-(defn custom
-  "Returns a validation function that checks pred for every item in the collection at key k.
-
-pred can be any function that yields a boolean. It will be invoked for every item in the collection.
-
-If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
-
-If no message is given, a default message will be used"
-  ([k pred]
-     (custom k pred :message (format "Custom validation failed for %s" (key->name k))))
-  ([k pred & {message :message}]
-     (mk-validator pred k message)))
-
-(defn every
-  "Returns a validation function that checks pred for every item in the collection at key k.
-
-pred can be any function that yields a boolean. It will be invoked for every item in the collection.
-
-If k is a vector, it is assumed to be the path in an nested associative structure. In this case, the :errors entry will mirror this path
-
-If no message is given, a default message will be used"
-  ([k pred]
-     (every k pred :message (format "All items in %s must satisfy the predicate" (key->name k))))
-  ([k pred & {message :message}]
-     (mk-validator #(every? pred %) k message)))
-
+;; ## Composability
 
 (defmacro defvalidatorset
   "Defines a set of validators encapsulating a reusable validation unit.
@@ -171,6 +168,7 @@ If no message is given, a default message will be used"
   forms should follow the semantics of \"bouncer.core/validate\"
 
   e.g.:
+
     (defvalidatorset addr-validator-set
       :postcode  [v/required v/number]
       :street    v/required